Fixed pseudocode.js and updated the bubble-sort plots
I switched to putting the pseudocode.js HTML in the math_helper.tmpl
template so it should work in the feed now. I had also forgotten that
the plots take place on the server where the jupyter kernel is running
so I copied them over to the nikola repository to be rendered in the
bubble sort post.